## Performance Improvements

- Added Benchmarks project to help measure performance improvements.
- Added dictionary to FieldMetaData to speed up metadata lookup.
- Moved tokenizer StringBuffer allocation to reduce allocations.
- Added configurable row strategy to tune performance.

## Convenience Features

- Changed FieldMetaData IndexOf to return -1 on not found instead of throwing TermNotFound exception.
- Added Length property to IFieldMetaData.
- Added new FactoryConfiguration class to provide configurability to ArchiveReader composite classes. This includes the following configuration classes for tuning performance.
  - FileReaderConfiguration
    - BufferSize - Allows tuning the size of the FileReader buffer.
  - RowFactoryConfiguration
    - Strategy - Allows selection of RowFactory strategy as follows:
      - Lazy - Uses IEnumerable.ElementAt to access fields. Use this strategy when accessing fields in a row in index order or when only accessing a subset of fields.
      - Greedy - Tokenizes all fields and converts them to an array. Use this strategy when accessing all fields in s row in random or reverse order.
  - ArchiveFolderConfiguration
    - OutputPath - Allows user to specify where archives are unzipped to.
    - ShouldCleanup - Allows user to control whether the folder is cleaned up after ArchiveReader is disposed.
- All terms are now readonly.
- Default values are now returned when they are defined. This also fixes a potential bug where default values are defined without an index.
- Restored support for netstandard2.0 and .NET Framework. If you are using version 0.2.0 then you can skip up to this version. Note: the netstandard2.0 version does not support IEnumerableAsync readers.

+++ src/DWC_A/ArchiveFolder.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using DwC_A.Config;
+using System;
 using System.IO;
 using System.IO.Compression;
 
@@ -8,18 +9,19 @@ namespace DwC_A
     {
         private readonly string fileName;
         private readonly string folderPath;
-        
+
         public bool ShouldCleanup { get; private set; }
 
         /// <summary>
         /// Extracts archive to a folder
         /// </summary>
         /// <param name="fileName">Zip archive file name</param>
-        /// <param name="folderPath">Path to extract to.  Leave null for a temp folder</param>
-        public ArchiveFolder(string fileName, string folderPath = null)
+        /// <param name="config"><see cref="ArchiveFolderConfiguration"/></param>
+        public ArchiveFolder(string fileName, ArchiveFolderConfiguration config)
         {
             this.fileName = fileName;
-            this.folderPath = string.IsNullOrEmpty(folderPath) ? GetTempPath() : folderPath;
+            ShouldCleanup = config.ShouldCleanup;
+            this.folderPath = string.IsNullOrEmpty(config.OutputPath) ? GetTempPath() : config.OutputPath;
         }
 
         public string Extract()
@@ -30,7 +32,10 @@ namespace DwC_A
 
         public void DeleteFolder()
         {
-            Directory.Delete(folderPath, true);
+            if (ShouldCleanup)
+            {
+                Directory.Delete(folderPath, true);
+            }
         }
 
         private string GetTempPath()
